+++ spring-blog/src/main/java/com/example/springblog/controller/AuthController.java
@@ -1,5 +1,6 @@
 package com.example.springblog.controller;
 
+import com.example.springblog.dto.LoginRequest;
 import com.example.springblog.dto.RegisterRequest;
 import com.example.springblog.service.AuthService;
 import org.springframework.http.HttpStatus;
@@ -23,5 +24,10 @@ public class AuthController {
     public ResponseEntity<Void> signup(@RequestBody RegisterRequest registerRequest) {
         authService.signup(registerRequest);
         return new ResponseEntity<>(HttpStatus.OK);
+    }
+
+    @PostMapping("/login")
+    public String login(@RequestBody LoginRequest loginRequest) {
+        return authService.login(loginRequest);
     }
 }